% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.baseline.r
\name{auto.baseline}
\alias{auto.baseline}
\title{Automatic Thermogram Baseline Subtraction}
\usage{
auto.baseline(
  x,
  w = 90,
  exclusion.lwr = 60,
  exclusion.upr = 80,
  grid.temp = seq(45, 90, 0.1),
  plot.on = FALSE,
  point = "innermost",
  explicit = FALSE
)
}
\arguments{
\item{x}{Raw thermogram data}

\item{w}{The number of points in the endpoint selection window. The number of points in the window. To find the lowest variance, "windows" of w points are selected from lowest temp to exclusion temp and the variance is calculated.}

\item{exclusion.lwr}{The lower bound of the exclusion window}

\item{exclusion.upr}{The upper bound of the exclusion window}

\item{grid.temp}{The set of temperatures to interpolate onto}

\item{plot.on}{logical: Should graphs be produced?}

\item{point}{The method of selecting the endpoint. Options are "innermost", "outmost", "mid"}

\item{explicit}{logical: Should text be displayed as the function runs}
}
\value{
Data frame with temperature and dCp of baseline subtracted and interpolated sample
}
\description{
Automatically detects two endpoints for baseline subtraction, then preforms baseline subtraction and returns final sample interpolated on the given grid
}
