% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoBaseline.R
\docType{data}
\name{DemoBaseline}
\alias{DemoBaseline}
\title{Demonstration Thermogram Data}
\format{
A data frame with 63,306 observations with 3 columns.
\describe{
\item{SampleID}{A unique identifier for each sample and iteration}
\item{Temperature}{The temperature measured in degrees celsius}
\item{dCp}{The dCp at the given temperature}
}
}
\usage{
DemoBaseline
}
\description{
A data frame containing demonstration thermograms in "long" format.
}
\keyword{datasets}
