% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.auto.baseline.r
\name{multiple.auto.baseline}
\alias{multiple.auto.baseline}
\title{Interpolates multiple thermograms}
\usage{
multiple.auto.baseline(
  x,
  w = 90,
  exclusion.lwr = 60,
  exclusion.upr = 80,
  grid.temp = seq(45, 90, 0.1),
  point = "outmost",
  explicit = TRUE,
  file.on = FALSE
)
}
\arguments{
\item{x}{Data frame of raw thermogram data. Must be filtered to desired grid beforehand}

\item{w}{The number of points in the window. To find the lowest variance, "windows" of w points are selected from lowest temp to exclusion temp and the variance is calculated.}

\item{exclusion.lwr}{The lower bound of the exclusion window}

\item{exclusion.upr}{The upper bound of the exclusion window}

\item{grid.temp}{The grid of temperatures the samples are interpolated onto}

\item{point}{The method of selecting the endpoint. Options are "innermost", "outmost", "mid".}

\item{explicit}{logical: Should text be displayed as the function runs}

\item{file.on}{logical: Should a csv file be saved of the output}
}
\value{
Data frame of temperature and dCp for multiple thermograms.
}
\description{
Takes in a data frame of multiple thermograms and returns a baseline subtracted data frame of clean thermograms
}
