% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.thermogram.subtraction.r
\name{multiple.thermogram.subtraction}
\alias{multiple.thermogram.subtraction}
\title{Intepolates multiple thermograms}
\usage{
multiple.thermogram.subtraction(
  x,
  w = 90,
  exclusion.lwr = 60,
  exclusion.upr = 80,
  grid.temp = seq(45, 90, 0.1),
  plot.on = FALSE,
  point = "outmost",
  explicit = TRUE,
  file.on = FALSE
)
}
\arguments{
\item{x}{Data frame of raw thermogram data. Must be filtered to desired grid beforehand}

\item{grid.temp}{The grid of temperatures the samples are interpolated onto}

\item{plot.on}{logical: should the output be graphed?}
}
\value{
Data frame of temperature and dCp
}
\description{
Fits a spline to the baseline subtracted data and interpolates the thermogram on the given temperature grid.
}
