% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.detection.r
\name{signal.detection}
\alias{signal.detection}
\title{Determines if the thermogram is signal or noise}
\usage{
signal.detection(sample)
}
\arguments{
\item{sample}{One thermogram sample with two columns labeled as Temperature and dCp}
}
\value{
Data frame with result of signal or no signal
}
\description{
Takes the first difference of the thermogram sample and uses auto.arima to determine if the sample is noise or if it is signal.
}
